change sample to .net core console app, change test project to .net core xunit project, create BuildAndTest.cmd
