+++ TT.StockQuoteSource.sln
@@ -9,13 +9,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{C9F3B1CE-113
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{6082211C-6941-43D2-9B52-68835EF8623A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TT.StockQuoteSource.Tests", "test\TT.StockQuoteSource.Tests\TT.StockQuoteSource.Tests.csproj", "{86830029-A949-41B8-932F-2BF13F1E3EDB}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TT.StockQuoteSource.Contracts", "src\TT.StockQuoteSource.Contract\TT.StockQuoteSource.Contracts.csproj", "{26F01B34-8029-4535-8FF2-D542A525EF7A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sample", "sample", "{FD4A04D2-275A-4DD7-ADB3-A0274D784766}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleApp", "sample\SampleApp.csproj", "{01FDCABE-D6E5-497B-A0BC-C1CB082DC7E5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TT.StockQuoteSource.TestsInNetCore", "test\TT.StockQuoteSource.TestsInNetCore\TT.StockQuoteSource.TestsInNetCore.csproj", "{3715E3D8-0024-4BFC-95A1-8639858A7A89}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SampleAppInNetCore", "sample\SampleAppInNetCore\SampleAppInNetCore.csproj", "{B1607EEE-81A0-4F31-8E6A-52DD73ECEFFD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,27 +27,27 @@ Global
 		{549F031A-FC19-46F6-8B06-916A9E647BDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{549F031A-FC19-46F6-8B06-916A9E647BDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{549F031A-FC19-46F6-8B06-916A9E647BDA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{86830029-A949-41B8-932F-2BF13F1E3EDB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{86830029-A949-41B8-932F-2BF13F1E3EDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{86830029-A949-41B8-932F-2BF13F1E3EDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{86830029-A949-41B8-932F-2BF13F1E3EDB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{26F01B34-8029-4535-8FF2-D542A525EF7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{26F01B34-8029-4535-8FF2-D542A525EF7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{26F01B34-8029-4535-8FF2-D542A525EF7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{26F01B34-8029-4535-8FF2-D542A525EF7A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{01FDCABE-D6E5-497B-A0BC-C1CB082DC7E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{01FDCABE-D6E5-497B-A0BC-C1CB082DC7E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{01FDCABE-D6E5-497B-A0BC-C1CB082DC7E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{01FDCABE-D6E5-497B-A0BC-C1CB082DC7E5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3715E3D8-0024-4BFC-95A1-8639858A7A89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3715E3D8-0024-4BFC-95A1-8639858A7A89}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3715E3D8-0024-4BFC-95A1-8639858A7A89}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3715E3D8-0024-4BFC-95A1-8639858A7A89}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B1607EEE-81A0-4F31-8E6A-52DD73ECEFFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1607EEE-81A0-4F31-8E6A-52DD73ECEFFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1607EEE-81A0-4F31-8E6A-52DD73ECEFFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1607EEE-81A0-4F31-8E6A-52DD73ECEFFD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{549F031A-FC19-46F6-8B06-916A9E647BDA} = {C9F3B1CE-1135-4966-85DF-8B0C7B406038}
-		{86830029-A949-41B8-932F-2BF13F1E3EDB} = {6082211C-6941-43D2-9B52-68835EF8623A}
 		{26F01B34-8029-4535-8FF2-D542A525EF7A} = {C9F3B1CE-1135-4966-85DF-8B0C7B406038}
-		{01FDCABE-D6E5-497B-A0BC-C1CB082DC7E5} = {FD4A04D2-275A-4DD7-ADB3-A0274D784766}
+		{3715E3D8-0024-4BFC-95A1-8639858A7A89} = {6082211C-6941-43D2-9B52-68835EF8623A}
+		{B1607EEE-81A0-4F31-8E6A-52DD73ECEFFD} = {FD4A04D2-275A-4DD7-ADB3-A0274D784766}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1866FEE8-BA9E-4E9F-BC48-79E0DBD832A8}
