--- sample/packages.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Extensions.Configuration" version="2.0.1" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Configuration.Abstractions" version="2.0.1" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Configuration.FileExtensions" version="2.0.1" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Configuration.Json" version="2.0.1" targetFramework="net461" />
-  <package id="Microsoft.Extensions.FileProviders.Abstractions" version="2.0.1" targetFramework="net461" />
-  <package id="Microsoft.Extensions.FileProviders.Physical" version="2.0.1" targetFramework="net461" />
-  <package id="Microsoft.Extensions.FileSystemGlobbing" version="2.0.1" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Primitives" version="2.0.0" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="10.0.1" targetFramework="net461" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.4.0" targetFramework="net461" />
-</packages>
