+++ src/TT.StockQuoteSource.Contract/StockQuoteSource.cs
@@ -5,8 +5,19 @@
     /// </summary>
     public enum StockQuoteSource
     {
+        /// <summary>
+        /// For testing
+        /// </summary>
         Test = -1,
+
+        /// <summary>
+        /// Yahoo Finance
+        /// </summary>
         Yahoo = 0,
+
+        /// <summary>
+        /// Alpha Vantage
+        /// </summary>
         AlphaVantage = 1,
     }
 }
