+++ test/TT.StockQuoteSource.TestsInNetCore/TT.StockQuoteSource.TestsInNetCore.csproj
@@ -1,0 +1,54 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Remove="TestFiles\AlphaVantage-HDV-1day-FullOutput.txt" />
+    <None Remove="TestFiles\AlphaVantage-HDV-1day.txt" />
+    <None Remove="TestFiles\AlphaVantage-HDV-5min.txt" />
+    <None Remove="TestFiles\YahooHtml-DRV.txt" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Include="TestFiles\AlphaVantage-HDV-1day-FullOutput.txt">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="TestFiles\AlphaVantage-HDV-1day.txt">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="TestFiles\AlphaVantage-HDV-5min.txt">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="TestFiles\YahooHtml-DRV.txt">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.0.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.7.0-preview-20180320-02" />
+    <PackageReference Include="xunit" Version="2.3.1" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
+    <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\TT.StockQuoteSource.Contract\TT.StockQuoteSource.Contracts.csproj" />
+    <ProjectReference Include="..\..\src\TT.StockQuoteSource\TT.StockQuoteSource.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="TestFiles\StockQuoteSourceConfig.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Update="TestFiles\YahooHistoricalData-2002.csv">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
+</Project>
