+++ test/TT.StockQuoteSource.TestsInNetCore/TestFiles/StockQuoteSourceConfig.json
@@ -1,0 +1,8 @@
+﻿{
+  "YahooFinanceURL": "https://finance.yahoo.com/quote/{0}",
+  "YahooHistoricalDataUrl": "https://query1.finance.yahoo.com/v7/finance/download/{0}?period1={1}&period2={2}&interval=1d&events=history&crumb={3}",
+  "AlphaVantageRealTimeUrl": "https://www.alphavantage.co/query?function=TIME_SERIES_INTRADAY&symbol={0}&interval=5min&apikey={1}",
+  "AlphaVantageDailyUrl": "https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol={0}&apikey={1}",
+  "AlphaVantageDailyFullOutputUrl": "https://www.alphavantage.co/query?function=TIME_SERIES_DAILY&symbol={0}&apikey={1}&outputsize=full",
+  "AlphaVantageKey": "{please-appy-your-own-key-at-alphavantage-website}"
+}
