+++ test/TT.StockQuoteSource.TestsInNetCore/TestFiles/YahooHistoricalData-2002.csv
@@ -1,0 +1,20 @@
+﻿Date,Open,High,Low,Close,Adj Close,Volume
+2018-03-09,23.500000,23.650000,23.500000,23.600000,23.600000,14120354
+2018-03-12,23.750000,23.799999,23.600000,23.700001,23.700001,14223731
+2018-03-13,23.850000,23.900000,23.650000,23.850000,23.850000,24457365
+2018-03-14,23.900000,23.900000,23.750000,23.850000,23.850000,10458025
+2018-03-15,23.850000,23.900000,23.750000,23.750000,23.750000,10181685
+2018-03-16,23.799999,23.900000,23.700001,23.700001,23.700001,34241985
+2018-03-19,23.700001,23.900000,23.700001,23.900000,23.900000,7397994
+2018-03-20,23.799999,23.850000,23.700001,23.750000,23.750000,11190178
+2018-03-21,23.799999,23.850000,23.750000,23.750000,23.750000,9817148
+2018-03-22,23.799999,23.900000,23.750000,23.850000,23.850000,10936647
+2018-03-23,23.650000,23.650000,23.450001,23.500000,23.500000,21586642
+2018-03-26,23.400000,23.450001,23.299999,23.450001,23.450001,13701322
+2018-03-27,23.500000,23.600000,23.500000,23.549999,23.549999,10481252
+2018-03-28,23.549999,23.549999,23.400000,23.400000,23.400000,7084591
+2018-03-29,23.400000,23.450001,23.200001,23.200001,23.200001,19957874
+2018-03-30,23.350000,23.549999,23.299999,23.500000,23.500000,11548942
+2018-04-02,23.500000,23.650000,23.450001,23.500000,23.500000,8901255
+2018-04-03,23.450001,23.549999,23.350000,23.350000,23.350000,14009186
+2018-04-09,23.350000,23.350000,23.350000,23.350000,23.350000,1
