--- tools/xunit.runner.console/xunit.console.exe.config
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-  <configSections>
-    <section name="xunit" type="Xunit.ConsoleClient.XunitConsoleConfigurationSection, xunit.console"/>
-  </configSections>
-
-  <xunit>
-    <transforms>
-      <add commandline="xmlv1" xslfile="xUnit1.xslt" description="output results to xUnit.net v1 style XML file"/>
-      <add commandline="nunit" xslfile="NUnitXml.xslt" description="output results to NUnit-style XML file"/>
-      <add commandline="html" xslfile="HTML.xslt" description="output results to HTML file"/>
-    </transforms>
-  </xunit>
-
-  <startup>
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/>
-  </startup>
-</configuration>
